test/MarkovChainTest.java

import org.junit.*;
import static org.junit.Assert.*;
import java.util.Arrays;

public class MarkovChainTest {

    @Test
    public void testAddBigram() {
        MarkovChain mc = new MarkovChain();
        mc.addBigram("1", "2");
        assertTrue(mc.chain.containsKey("1"));
        ProbabilityDistribution<String> pd = mc.chain.get("1");
        assertTrue(pd.getRecords().containsKey("2"));
        assertEquals(1, pd.count("2"));
    }

    @Test
    public void testTrain() {
        MarkovChain mc = new MarkovChain();
        String sentence = "1 2 3";
        mc.train(Arrays.stream(sentence.split(" ")).iterator());
        assertEquals(3, mc.chain.size());
        ProbabilityDistribution<String> pd1 = mc.chain.get("1");
        assertTrue(pd1.getRecords().containsKey("2"));
        assertEquals(1, pd1.count("2"));
        ProbabilityDistribution<String> pd2 = mc.chain.get("2");
        assertTrue(pd2.getRecords().containsKey("3"));
        assertEquals(1, pd2.count("3"));
        ProbabilityDistribution<String> pd3 = mc.chain.get("3");
        assertTrue(pd3.getRecords().containsKey(null));
        assertEquals(1, pd3.count(null));
    }

    @Test
    public void testWalk() {

        Integer[] walkIndices = {0, 0, 0, 1, 1, 0};
	    String[] words = {"CIS", "120", "beats", "CIS", "120", "rocks"};
        NumberGenerator r = new ListNumberGenerator(walkIndices);
        MarkovChain mc = new MarkovChain(r);
        
        String sentence1 = "CIS 120 rocks";
        String sentence2 = "CIS 120 beats CIS 160";
        mc.train(Arrays.stream(sentence1.split(" ")).iterator());
        mc.train(Arrays.stream(sentence2.split(" ")).iterator());
        
        mc.reset("CIS"); 
        for (int i = 0; i < words.length; i++) {
        	assertTrue(mc.hasNext());
            assertEquals(words[i], mc.next());
        }
    }
}
